0-read_textfile.c

+++ 0x15-file_io/0-read_textfile.c
@@ -11,7 +11,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int fd = 0, bytes = 0, nbytes = 0;
+	ssize_t fd = 0, bytes = 0, nbytes = 0;
 	char *buf;
 
 	if (filename == NULL)
